refactor: move updated skeleton forest creation to be under updated skeleton forest
